requirements.txt
aiogram==3.18.0
tortoise-orm==0.24.2
aiosqlite==0.21.0
APScheduler==3.11.0

gspread==6.2.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.164.0
google-api-core==2.24.2
googleapis-common-protos==1.69.2
httplib2==0.22.0
proto-plus==1.26.1
protobuf==6.30.1
oauthlib==3.2.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
uritemplate==4.1.1
oauth2client
openai==1.66.5

aiohttp==3.11.14
aiofiles==24.1.0
requests==2.32.3
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
urllib3==2.3.0

pyTelegramBotAPI

pytest
pytest-mock
pytest-asyncio
tenacity
gspread-formatting